requirements.txt
colour == 0.1.4
BlinkStick == 1.2.0
PyYAML == 3.12
